changed SelectedNode to an Int

+++ src/main/kotlin/org/team2471/frc/nodeDeck/NodeSelector.kt
@@ -18,7 +18,7 @@ object NodeSelector : GridPane() {
     val nine = Button("")
     val buttonSideLength: Double = 400.0
     var selectedNodeButton: Button = one
-    var selectedNode: Double = 1.0
+    var selectedNode: Int = 1
 
 
     init {
@@ -28,8 +28,8 @@ object NodeSelector : GridPane() {
         one.style = "-fx-background-color: #FFFF00"
         one.setOnMousePressed {
             if (TableCreator.isRedAlliance) {
-                selectedNode = 1.0
-            } else selectedNode = 3.0
+                selectedNode = 1
+            } else selectedNode = 3
             changeSelectedNodeButton(one)
             InformationPanel.updateInfoPanel()
         }
@@ -37,7 +37,7 @@ object NodeSelector : GridPane() {
         two.setPrefSize(buttonSideLength, buttonSideLength)
         two.style = "-fx-background-color: #9900ff"
         two.setOnMousePressed {
-            selectedNode = 2.0
+            selectedNode = 2
             changeSelectedNodeButton(two)
             InformationPanel.updateInfoPanel()
         }
@@ -46,8 +46,8 @@ object NodeSelector : GridPane() {
         three.style = "-fx-background-color: #FFFF00"
         three.setOnMousePressed {
             if (TableCreator.isRedAlliance) {
-                selectedNode = 3.0
-            } else selectedNode = 1.0
+                selectedNode = 3
+            } else selectedNode = 1
             changeSelectedNodeButton(three)
             InformationPanel.updateInfoPanel()
         }
@@ -56,8 +56,8 @@ object NodeSelector : GridPane() {
         four.style = "-fx-background-color: #FFFF00"
         four.setOnMousePressed {
             if (TableCreator.isRedAlliance) {
-                selectedNode = 4.0
-            } else selectedNode = 6.0
+                selectedNode = 4
+            } else selectedNode = 6
             changeSelectedNodeButton(four)
             InformationPanel.updateInfoPanel()
         }
@@ -65,7 +65,7 @@ object NodeSelector : GridPane() {
         five.setPrefSize(buttonSideLength, buttonSideLength)
         five.style = "-fx-background-color: #9900ff"
         five.setOnMousePressed {
-            selectedNode = 5.0
+            selectedNode = 5
             changeSelectedNodeButton(five)
             InformationPanel.updateInfoPanel()
         }
@@ -74,8 +74,8 @@ object NodeSelector : GridPane() {
         six.style = "-fx-background-color: #FFFF00"
         six.setOnMousePressed {
             if (TableCreator.isRedAlliance) {
-                selectedNode = 6.0
-            } else selectedNode = 4.0
+                selectedNode = 6
+            } else selectedNode = 4
             changeSelectedNodeButton(six)
             InformationPanel.updateInfoPanel()
         }
@@ -84,8 +84,8 @@ object NodeSelector : GridPane() {
         seven.style = "-fx-background-color: #595959"
         seven.setOnMousePressed {
             if (TableCreator.isRedAlliance) {
-                selectedNode = 7.0
-            } else selectedNode = 9.0
+                selectedNode = 7
+            } else selectedNode = 9
             changeSelectedNodeButton(seven)
             InformationPanel.updateInfoPanel()
         }
@@ -93,7 +93,7 @@ object NodeSelector : GridPane() {
         eight.setPrefSize(buttonSideLength, buttonSideLength / 2 + 33.7)
         eight.style = "-fx-background-color: #595959"
         eight.setOnMousePressed {
-            selectedNode = 8.0
+            selectedNode = 8
             changeSelectedNodeButton(eight)
             InformationPanel.updateInfoPanel()
         }
@@ -102,8 +102,8 @@ object NodeSelector : GridPane() {
         nine.style = "-fx-background-color: #595959"
         nine.setOnMousePressed {
             if (TableCreator.isRedAlliance) {
-                selectedNode = 9.0
-            } else selectedNode = 7.0
+                selectedNode = 9
+            } else selectedNode = 7
             changeSelectedNodeButton(nine)
             InformationPanel.updateInfoPanel()
         }
